Moved clientsocket to network in App since that makes more sense
