+++ src/main/java/no/ntnu/IDATA2304/Group1/clientapp/app2/network/ClientSocket.java
