+++ src/test/resources/no/ntnu/idata2304/group1/clientapp/app2/network/trustedCerts/Gruppe_1.cer
